move name generator to utils
