+++ src/components/clip.tsx
@@ -38,7 +38,7 @@ import generateId from '../utils/generateId';
 import determineOpacity from '../utils/determineOpacity';
 import { Point } from '@turf/helpers';
 import { clipSteps } from '../tutorial/steps/clipSteps';
-import Tutorial from '../tutorial/steps/tutorial';
+import Tutorial from '../tutorial/tutorial';
 
 const ITEM_HEIGHT = 48;
 const ITEM_PADDING_TOP = 8;
