+++ src/components/union.tsx
@@ -12,10 +12,11 @@ import { modalStyle } from './styledComponents';
 import processData from '../utils/flattenAndDissolve';
 import { generateColor } from '../utils/genereateColor';
 import generateId from '../utils/generateId';
+import createUniqueName from '../utils/createUniqueName';
 import InfoIcon from '@mui/icons-material/Info';
 import { unionSteps } from '../tutorial/steps/unionSteps';
 import makeStyles from '@mui/styles/makeStyles';
-import Tutorial from '../tutorial/steps/tutorial';
+import Tutorial from '../tutorial/tutorial';
 
 const useStyles = makeStyles({
   hovered: {
@@ -87,7 +88,7 @@ function Union(props: {
         if (unioned?.features.length > 0) {
           const newObj: GeoJSONItem = {
             id: generateId(),
-            name: createUniqueName(name),
+            name: createUniqueName(name, geoJSONList),
             visible: true,
             color: generateColor(),
             opacity: 0.5,
@@ -142,18 +143,6 @@ function Union(props: {
       props.showAlert('warning', 'Please select a polygon layer');
     }
   };
-
-  function createUniqueName(name: string) {
-    let count = 0;
-    const baseName = name;
-    const names = geoJSONList.map((item) => item.name);
-
-    while (names.includes(name)) {
-      count++;
-      name = `${baseName}_${count}`;
-    }
-    return name;
-  }
 
   return (
     <>
